requirements.txt
opencv-python
torch
torchvision
transformers
numpy
matplotlib
scikit-learn
Pillow
tqdm
streamlit
sentencepiece
protobuf
